+++ CHANGELOG.md
@@ -6,7 +6,9 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 ## dev
 
 * Enhancements
-  * [Manifest] `wait` accepts a simple number #148
+  * [Cli] Start github URLs directly on `azk start` #476;
+  * [Manifest] `wait` accepts a simple number #148;
+  * [Package] `git` is an `azk` installation dependency #476;
 
 ## v0.14.4 - (2015-06-26)